requirements.txt
streamlit
pytesseract
Pillow
textblob
openpyxl
pandas


https://github.com/tesseract-ocr/tessdata/tree/main